--- reporters/kamon-opentelemetry-http/src/main/resources/reference.conf
@@ -1,46 +1,0 @@
-# ======================================== #
-# kamon-otlp-http reference configuration  #
-# ======================================== #
-
-kamon.otel-http.trace {
-  # Hostname and port where the OTLP Server is running
-  host = "localhost"
-  port = 4318
-
-  # Decides whether to use HTTP or HTTPS when connecting to the OTel server
-  schema = "http"
-
-  # default to support the ENV:s as described at
-  # https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/protocol/exporter.md
-  # Must be full url path (e.g. must include /v1/traces if forwarding to an otel collector with default path configuration)
-  endpoint = ${kamon.otel-http.trace.schema}"://"${kamon.otel-http.trace.host}":"${kamon.otel-http.trace.port}"/v1/traces"
-  endpoint = ${?OTEL_EXPORTER_OTLP_ENDPOINT}
-  endpoint = ${?OTEL_EXPORTER_OTLP_TRACES_ENDPOINT}
-  # For supporting the unlikely scenario where both grpc and http otel reporters are enabled
-  endpoint = ${?OTEL_EXPORTER_OTLP_HTTP_TRACES_ENDPOINT}
-
-  # Enable or disable including tags from kamon.environment.tags as resource labels in the exported trace
-  # Any keys containing a hyphen (-) will be converted to a dot (.) as that is the standard.
-  # e.g. service-version becomes service.version
-  # reference: https://github.com/kamon-io/Kamon/blob/master/core/kamon-core/src/main/resources/reference.conf#L23
-  include-environment-tags = no
-}
-
-# Arbitrary key-value pairs that further identify the environment where this service instance is running.
-# These are added as KeyValue labels to the Resource part of the exported traces
-# Requires 'include-environment-tags' to be set to 'yes'
-#
-# kamon.environment.tags {
-#   service-version = "x.x.x"
-#   service-namespace = "ns"
-#   service-instance.id = "xxx-yyy"
-# }
-
-kamon.modules {
-  otel-http-trace-reporter {
-    enabled = true
-    name = "OpenTelemetry Trace Reporter"
-    description = "Sends trace data to a OpenTelemetry server via http/protobuf"
-    factory = "kamon.otel_http.OpenTelemetryHttpTraceReporter$Factory"
-  }
-}
