--- reporters/kamon-opentelemetry-http/src/test/resources/application.conf
@@ -1,22 +1,0 @@
-kamon {
-  #force sampling (1 = 100%), normally nothing one sets in the config but used for this demo purpose
-  #setting is from kamon
-  trace.random-sampler.probability = 1
-
-  #configure the OTel exporter
-  otel.trace {
-    include-environment-tags = yes
-  }
-
-  environment {
-    service = "kamon-test-application"
-
-    # These will be added as Resource labels in the trace export
-#    tags {
-#      service-version = "x.x.x"
-#      service-instance-id = "GERF43253FV"
-#      env = "kamon-devint"
-#    }
-
-  }
-}
