--- reporters/kamon-opentelemetry/src/main/resources/META-INF/native-image/io.kamon/kamon-opentelemetry/native-image.properties
@@ -1,1 +1,0 @@
-Args = -H:IncludeResources=.*\\.conf -H:ReflectionConfigurationResources=${.}/reflect-config.json
