--- reporters/kamon-opentelemetry/src/main/resources/META-INF/native-image/io.kamon/kamon-opentelemetry/reflect-config.json
@@ -1,6 +1,0 @@
-[
-  {
-    "name": "kamon.otel.OpenTelemetryTraceReporter$Factory",
-    "allDeclaredConstructors": true
-  }
-]
