+++ reporters/kamon-opentelemetry/src/main/scala/kamon/otel/SpanConverter.scala
@@ -15,212 +15,110 @@
  */
 package kamon.otel
 
-import com.google.protobuf.ByteString
-import io.opentelemetry.proto.common.v1.{AnyValue, InstrumentationLibrary, KeyValue}
-import io.opentelemetry.proto.resource.v1.Resource
-import io.opentelemetry.proto.trace.v1.{InstrumentationLibrarySpans, ResourceSpans, Status, Span => ProtoSpan}
-import kamon.tag.{Lookups, Tag}
-import kamon.tag.Tag.unwrapValue
-import kamon.trace.Span.{Kind, TagKeys}
-import kamon.trace.{Identifier, Span}
-import org.slf4j.LoggerFactory
-
 import java.time.Instant
-import java.util.Collections
+import java.util
 import java.util.concurrent.TimeUnit
+import java.util.{Collection => JCollection}
+
+import scala.collection.JavaConverters._
+
+import io.opentelemetry.api.common.Attributes
+import io.opentelemetry.api.trace._
+import io.opentelemetry.sdk.common.InstrumentationLibraryInfo
+import io.opentelemetry.sdk.resources.Resource
+import io.opentelemetry.sdk.trace.data.{EventData, LinkData, SpanData, StatusData}
+import kamon.tag.{Lookups, Tag, TagSet}
+import kamon.trace.Span.{Kind, TagKeys}
+import kamon.trace.Trace.SamplingDecision.Sample
+import kamon.trace.{Identifier, Span}
+
+
+class SpanWrapper(resource: Resource, instrumentationLibraryInfo: InstrumentationLibraryInfo, span: Span.Finished) extends SpanData {
+  private val sampled: Boolean = span.trace.samplingDecision == Sample
+
+  override def getName: String = span.operationName
+
+  override val getKind: SpanKind = SpanConverter.toKind(span.kind)
+
+  override val getSpanContext: SpanContext = SpanConverter.mkSpanContext(sampled, span.trace.id, span.id)
+
+  override val getParentSpanContext: SpanContext =
+    if (span.parentId.isEmpty) SpanContext.getInvalid else SpanConverter.mkSpanContext(sampled, span.trace.id, span.parentId)
+
+  override val getStatus: StatusData = SpanConverter.getStatus(span)
+
+  override val getStartEpochNanos: Long = SpanConverter.toEpocNano(span.from)
+
+  override val getAttributes: Attributes = SpanConverter.toAttributes(span.tags)
+
+  override val getEvents: util.List[EventData] = span.marks.map(SpanConverter.toEvent).asJava
+
+  override val getLinks: util.List[LinkData] = span.links.map(SpanConverter.toLink).asJava
+
+  override val getEndEpochNanos: Long = SpanConverter.toEpocNano(span.to)
+
+  override def hasEnded: Boolean = true
+
+  override def getTotalRecordedEvents: Int = span.marks.size
+
+  override def getTotalRecordedLinks: Int = span.links.size
+
+  override def getTotalAttributeCount: Int = getAttributes.size()
+
+  override def getInstrumentationLibraryInfo: InstrumentationLibraryInfo = instrumentationLibraryInfo
+
+  override def getResource: Resource = resource
+}
 
 /**
- * Converts Kamon spans to OpenTelemetry proto Spans
- */
+  * Converts Kamon spans to OpenTelemetry [[SpanData]]s
+  */
 private[otel] object SpanConverter {
-  private val logger = LoggerFactory.getLogger(SpanConverter.getClass)
+  private[otel] def toEpocNano(instant: Instant): Long =
+    TimeUnit.NANOSECONDS.convert(instant.getEpochSecond, TimeUnit.SECONDS) + instant.getNano
 
-  /** 8-bytes worth of zeroes used to pad 8-byte TraceId's  */
-  private val eightBytePad = Array.fill[Byte](8)(0)
-
-  /** Creates a AnyValue type KeyValue */
-  private[otel] def anyKeyValue(key:String, value:AnyValue):KeyValue = KeyValue.newBuilder().setKey(key.replaceAll("-", ".")).setValue(value).build
-  /** Creates a String type KeyValue */
-  private[otel] def stringKeyValue(key:String, value:String):KeyValue = anyKeyValue(key, AnyValue.newBuilder().setStringValue(value).build())
-  /** Creates a Boolean type KeyValue */
-  private[otel] def booleanKeyValue(key:String, value:Boolean):KeyValue = anyKeyValue(key, AnyValue.newBuilder().setBoolValue(value).build())
-  /** Creates a Long type KeyValue */
-  private[otel] def longKeyValue(key:String, value:Long):KeyValue = anyKeyValue(key,AnyValue.newBuilder().setIntValue(value).build())
-
-  /**
-   * Converts a sequence of Kamon spans to a proto ''ResourceSpans''
-   * @param resource resource information for this instance of the service. Added as resource labels
-   * @param instrumentationLibrary instrumentation library information to add to the exported spans
-   * @param spans The spans to export
-   * @return
-   */
-  private[otel] def toProtoResourceSpan(resource:Resource, instrumentationLibrary:InstrumentationLibrary)(spans:Seq[Span.Finished]):ResourceSpans = {
-    import collection.JavaConverters._
-
-    //it is assumed all spans belong to the same instrumentation library
-    val instrumentationLibrarySpans = InstrumentationLibrarySpans.newBuilder()
-      .setInstrumentationLibrary(instrumentationLibrary)
-      .addAllSpans(spans.map(toProtoSpan).asJava)
-      .build()
-
-    ResourceSpans.newBuilder()
-      .setResource(resource)
-      .addAllInstrumentationLibrarySpans(Collections.singletonList(instrumentationLibrarySpans))
-      .build()
+  private[otel] def toKind(kind: Span.Kind): SpanKind = kind match {
+    case Kind.Client => SpanKind.CLIENT
+    case Kind.Consumer => SpanKind.CONSUMER
+    case Kind.Internal => SpanKind.INTERNAL
+    case Kind.Producer => SpanKind.PRODUCER
+    case Kind.Server => SpanKind.SERVER
+    case _ => SpanKind.INTERNAL // Default value
   }
 
-  /**
-   * Converts a Kamon span to a proto span
-   * @param span the span to convert
-   * @return
-   */
-  private[otel] def toProtoSpan(span:Span.Finished):ProtoSpan = {
-    //converts Kamon span tags to proto KeyValue attributes
-    val attributes:List[KeyValue] = span.tags.iterator().map(toProtoKeyValue).toList
+  private[otel] def getTraceFlags(sample: Boolean): TraceFlags =
+    if (sample) TraceFlags.getSampled else TraceFlags.getDefault
 
-    //converts Kamon span links to proto links
-    val links:Seq[ProtoSpan.Link] = span.links.map(toProtoLink)
-    //converts Kamon span marks to proto events
-    val events:Seq[ProtoSpan.Event] = span.marks.map(toProtoEvent)
+  private[otel] def mkSpanContext(sample: Boolean, traceId: Identifier, spanId: Identifier): SpanContext = {
+    val paddedTraceId = if (traceId.string.length == 16) s"0000000000000000${traceId.string}" else traceId.string
+    SpanContext.create(paddedTraceId, spanId.string, getTraceFlags(sample), TraceState.getDefault)
+  }
 
-    import collection.JavaConverters._
-    val builder = ProtoSpan.newBuilder()
-      .setTraceId(traceIdToByteString(span.trace.id))
-      .setSpanId(spanIdToByteString(span.id))
-      .setName(span.operationName)
-      .setKind(toProtoKind(span.kind))
-      .setStartTimeUnixNano(toEpocNano(span.from))
-      .setEndTimeUnixNano(toEpocNano(span.to))
-      .setStatus(getStatus(span))
-      .addAllAttributes(attributes.asJava)
-      .addAllLinks(links.asJava)
-      .addAllEvents(events.asJava)
+  private[otel] def getStatus(span: Span.Finished): StatusData =
+    if (span.hasError) {
+      val msg = span.tags.get(Lookups.option(TagKeys.ErrorMessage)) getOrElse ""
+      StatusData.create(StatusCode.ERROR, msg)
+    } else StatusData.ok()
 
-    //TODO add set traceState once we have something to set. Need w3c context
-    //It is a trace_state in w3c-trace-context format: https://www.w3.org/TR/trace-context/#tracestate-header
-    //See also https://github.com/w3c/distributed-tracing for more details about this field.
-    //builder.setTraceState()
+  private[otel] def toEvent(mark: Span.Mark): EventData =
+    EventData.create(toEpocNano(mark.instant), mark.key, Attributes.empty())
 
-    //add optional parent
-    val parentId = span.parentId
-    if(!parentId.isEmpty) builder.setParentSpanId(spanIdToByteString(parentId))
+  private[otel] def toLink(link: Span.Link): LinkData =
+    LinkData.create(mkSpanContext(link.trace.samplingDecision == Sample, link.trace.id, link.spanId))
 
+  private[otel] def toAttributes(tags: TagSet): Attributes = {
+    val builder = Attributes.builder()
+    tags.iterator().foreach {
+      case t: Tag.String => builder.put(t.key, t.value)
+      case t: Tag.Boolean => builder.put(t.key, t.value)
+      case t: Tag.Long => builder.put(t.key, t.value)
+    }
     builder.build()
   }
 
-  /**
-   * Converts the instant to EPOC nanos
-   * @param instant
-   * @return
-   */
-  private[otel] def toEpocNano(instant:Instant):Long = TimeUnit.NANOSECONDS.convert(instant.getEpochSecond, TimeUnit.SECONDS) + instant.getNano
+  def convertSpan(resource: Resource, instrumentationLibrary: InstrumentationLibraryInfo)(span: Span.Finished): SpanData =
+    new SpanWrapper(resource, instrumentationLibrary, span)
 
-  /**
-   * Creates a proto ''Status'' of the Kamon span status
-   * @param span
-   * @return
-   */
-  private[otel] def getStatus(span:Span.Finished):Status = {
-    //according to the spec the deprecated code needs to be set for backwards compatibility reasons
-    //https://github.com/open-telemetry/opentelemetry-proto/blob/main/opentelemetry/proto/trace/v1/trace.proto
-    val (status, deprecatedStatus) = if(span.hasError) (Status.StatusCode.STATUS_CODE_ERROR_VALUE, Status.DeprecatedStatusCode.DEPRECATED_STATUS_CODE_UNKNOWN_ERROR_VALUE) else (Status.StatusCode.STATUS_CODE_OK_VALUE, Status.DeprecatedStatusCode.DEPRECATED_STATUS_CODE_OK_VALUE)
-    val builder = Status.newBuilder()
-      .setCodeValue(status)
-      .setDeprecatedCodeValue(deprecatedStatus)
-
-    //if there is an error message in the span we add it as status message
-    //spec states: "A developer-facing human readable error message."
-    //https://github.com/open-telemetry/opentelemetry-proto/blob/main/opentelemetry/proto/trace/v1/trace.proto
-    span.tags.get(Lookups.option(TagKeys.ErrorMessage)).foreach(builder.setMessage)
-
-    builder.build()
-  }
-
-  /**
-   * Converts a Kamon span kind to a proto span kind
-   * @param kind
-   * @return
-   */
-  private[otel] def toProtoKind(kind:Kind): ProtoSpan.SpanKind = {
-    kind match {
-      case Kind.Client => ProtoSpan.SpanKind.SPAN_KIND_CLIENT
-      case Kind.Consumer => ProtoSpan.SpanKind.SPAN_KIND_CONSUMER
-      case Kind.Internal => ProtoSpan.SpanKind.SPAN_KIND_INTERNAL
-      case Kind.Producer => ProtoSpan.SpanKind.SPAN_KIND_PRODUCER
-      case Kind.Server => ProtoSpan.SpanKind.SPAN_KIND_SERVER
-      case Kind.Unknown => ProtoSpan.SpanKind.SPAN_KIND_UNSPECIFIED
-      case _ => ProtoSpan.SpanKind.UNRECOGNIZED
-    }
-  }
-
-  /**
-   * Converts a Kamon tag to a proto KeyValue
-   * @param tag
-   * @return
-   */
-  private[otel] def toProtoKeyValue(tag: Tag):KeyValue = {
-    tag match {
-      case t: Tag.String  => stringKeyValue(tag.key, t.value)
-      case t: Tag.Boolean => booleanKeyValue(tag.key, t.value)
-      case t: Tag.Long    => longKeyValue(tag.key, t.value)
-      case _ => { //this cannot happen unless new Tag types are added to Kamon core, the code is more of a safeguard for "just in case"
-        logger.warn(s"Failed to map Tag [$tag] to a KeyValue type, will attempt to convert to a string")
-        stringKeyValue(tag.key, unwrapValue(tag).toString) //last resort trying to create a string of this unknown Tag type
-      }
-    }
-  }
-
-  /**
-   * Converts a Kamon span link to a proto span link
-   * @param link
-   * @return
-   */
-  private[otel] def toProtoLink(link:Span.Link):ProtoSpan.Link = {
-    ProtoSpan.Link.newBuilder()
-      .setTraceId(traceIdToByteString(link.trace.id))
-      .setSpanId(spanIdToByteString(link.spanId))
-      .build()
-  }
-
-  /**
-    * Converts a Kamon Mark to a proto span event
-    * @param mark
-    * @return
-    */
-  private[otel] def toProtoEvent(mark:Span.Mark):ProtoSpan.Event = {
-    ProtoSpan.Event.newBuilder()
-      .setName(mark.key)
-      .setTimeUnixNano(toEpocNano(mark.instant))
-      .build()
-  }
-
-  /**
-   * Converts the TraceID to a proto ByteString.
-   * Id the identifier is 8-bytes (single) it will be padded to 16-bytes as this is required in OTLP
-   * @param id
-   * @return
-   */
-  private[otel] def traceIdToByteString(id:Identifier):ByteString = toByteString(id, true)
-
-  /**
-   * Converts the SpanID to a proto ByteString
-   * @param id
-   * @return
-   */
-  private[otel] def spanIdToByteString(id:Identifier):ByteString = toByteString(id, false)
-
-  /**
-   * Converts a Kamon identifier to a proto ByteString
-   * @param id The identifier to convert
-   * @param padTo16bytes If the identifier should be padded in case it is 8-bytes (used for traceid's)
-   * @return
-   */
-  private[otel] def toByteString(id:Identifier, padTo16bytes:Boolean):ByteString = {
-    val bytes = id.bytes
-    //OTLP requires 16-bytes trace identifiers, if we have 8-byte traceid it needs to padded with zeroes
-    if(padTo16bytes && id.bytes.length == 8)
-      ByteString.copyFrom(eightBytePad++bytes)
-    else
-      ByteString.copyFrom(bytes)
-  }
+  def convert(resource: Resource, instrumentationLibrary: InstrumentationLibraryInfo)(spans: Seq[Span.Finished]): JCollection[SpanData] =
+    spans.map(convertSpan(resource, instrumentationLibrary)).asJava
 }
