+++ reporters/kamon-opentelemetry/src/test/resources/application.conf
@@ -12,11 +12,11 @@ kamon {
     service = "kamon-test-application"
 
     # These will be added as Resource labels in the trace export
-    tags {
-      service-version = "x.x.x"
-      service-instance-id = "GERF43253FV"
-      env = "kamon-devint"
-    }
+#    tags {
+#      service-version = "x.x.x"
+#      service-instance-id = "GERF43253FV"
+#      env = "kamon-devint"
+#    }
 
   }
 }
